ReportActionCompose 4/6: extract input hooks
